Changes in the `mayFreeIoBlobs` criterion. (#1126)
First off all, it was rewritten as function to simplify understanding.
But this is not the main change.
The main is that `!IsRecurrentMode() || IsLastSequencePos()` subcriteron is combined
with others using `and` instead of `or`. As a result, the inputs of an incomplete
recurrent layer will not be dropped until the recurrent processing completes.

Found when monitoring allocations: each call of `runOnce` inside recursion was
allocating output blob.

Signed-off-by: slon872 <30573417+slon872@users.noreply.github.com>